Revert "Split 'plugins' and 'sidebar' keys out of `config.js` for easier handling"
This reverts commit b5bbfa1b4a4a77259906eb9f53e4c9bd2c3419c8.
